feat (multiboot_params): better structure
