+++ Sources/Kernel.cpp
@@ -8,64 +8,19 @@
 #error "This OS needs to be compiled with a x86_64-elf compiler"
 #endif
 
-#include "Boot/multiboot.h"
 #include "VgaBuffer.hpp"
+#include "Boot/multiboot.hpp"
 #include "Utils/Functions/PrintUtils.hpp"
 #include "Interrupts/InterruptsDescriptorTable.hpp"
 #include "Interrupts/InterruptHandlersGenerator/InterruptHandlersGenerator.hpp"
-
-void parse_multiboot_info(const uint32_t &magic, const uintptr_t &addr)
-{
-    if (magic != MULTIBOOT2_BOOTLOADER_MAGIC)
-    {
-        PrintUtils::printk("Invalid magic number: %x\n", magic);
-        THROW_KERNEL_EXCEPTION();
-    }
-
-    multiboot_info_t *mb_info = (multiboot_info_t *) addr;
-    uint8_t *ptr = (uint8_t *) &mb_info->tags[0];
-    while (1) {
-        struct multiboot_tag *tag = (struct multiboot_tag *) ptr;
-
-        if (tag->type == MULTIBOOT_TAG_TYPE_END) {
-            break;
-        }
-
-        switch (tag->type) {
-            case MULTIBOOT_TAG_TYPE_CMDLINE: {
-                struct multiboot_tag_string *cmd = (struct multiboot_tag_string *) tag;
-                PrintUtils::printk("Command line tag found, cmdline: %s\n", cmd->string);
-                // use cmd->string
-                break;
-            }
-            case MULTIBOOT_TAG_TYPE_MMAP: {
-                struct multiboot_tag_mmap *mmap = (struct multiboot_tag_mmap *) tag;
-                PrintUtils::printk("Memory map tag found, size: %d\n", mmap->size);
-                // walk mmap->entries[...] here
-                break;
-            }
-            case MULTIBOOT_TAG_TYPE_FRAMEBUFFER: {
-                PrintUtils::printk("Framebuffer info tag found!\n");
-                // handle framebuffer info if you requested it
-                break;
-            }
-            default:
-                PrintUtils::printk("Unknown tag type %d\n", tag->type);
-                // ignore unknown tag
-                break;
-        }
-
-        // advance to next tag (8-byte aligned)
-        ptr += (tag->size + 7) & ~7;
-    }
-}
 
 extern "C"
 {
     void kernel_main(uint32_t magic, uintptr_t addr)
 	{
         Terminal::get().initialize();
-        parse_multiboot_info(magic, addr);
+        Multiboot multiboot = Multiboot(magic, addr);
+        multiboot.parse_mb_info();
 
         InterruptHandlersGenerator interruptHandlersGenerator;
         InterruptDescriptorTable idt;
